requirements/prod.txt
python-dotenv
openai
fastapi
uvicorn
scikit-learn